Alot of testing, changed snek to arraylist and use trail like patern to keep track of previous locations => development branch next

+++ websocketshared/src/main/java/com/s3/SnekIO/websocketshared/models/InputMouse.java
@@ -3,27 +3,27 @@ package com.s3.SnekIO.websocketshared.models;
 import com.s3.SnekIO.websocketshared.actions.IAction;
 
 public class InputMouse implements IAction {
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
-    public InputMouse(int x, int y) {
+    public InputMouse(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
